+++ src/component/mypage/MyReview.js
@@ -1,52 +1,20 @@
-import React, { useEffect, useState } from "react";
+import React from "react";
 import InquireBoxTitle from "../utils/InquireBoxTitle";
 import Reserve from "../../img/list-check-solid.svg";
-import useApi from "../../hook/useApi";
 import HandleReview from "../book/review/HandleReview";
-// import axiosPromise from "../../util/axios";
+import axiosPromise from "../../util/axios";
 import Pagination from "../utils/Pagination";
-
-const useGetReviewInfo = page => {
-  const [lastPage, setLastPage] = useState(null);
-  const [reviewInfo, setreviewInfo] = useState(null);
-  const [reviewList, setReviewList] = useState([]);
-  const userId = JSON.parse(window.localStorage.getItem("user")).id;
-  const { request, Dialog } = useApi(
-    "get",
-    `reviews?userId=${userId}&page=${page - 1}`,
-  );
-
-  const refineResponse = response => {
-    setreviewInfo(response.data.meta);
-    setLastPage(response.data.meta.totalPages);
-    setReviewList(response.data.items);
-  };
-
-  return {
-    request,
-    lastPage,
-    reviewList,
-    reviewInfo,
-    refineResponse,
-    Dialog,
-  };
-};
+import useGetReviewInfo from "../../api/review/useGetReviewInfo";
 
 const MyReview = () => {
-  const [page, setPage] = useState(1);
-  const { request, lastPage, reviewList, refineResponse, Dialog } =
-    useGetReviewInfo(page);
-  console.log(page);
-  console.log(reviewList);
-  useEffect(() => {
-    request(refineResponse);
-  }, [page]);
-  // const deleteReview = reviewId => {
-  //   const temp = postReviews.filter(review => review.reviewsId !== reviewId);
-  //   setPostReviews(temp);
-  //   axiosPromise("delete", `/reviews/${reviewId}`);
-  // };
-  // console.log(page);
+  const { page, setPage, lastPage, reviewList, setReviewList, Dialog } =
+    useGetReviewInfo();
+
+  const deleteReview = reviewId => {
+    const temp = reviewList.filter(review => review.reviewsId !== reviewId);
+    setReviewList(temp);
+    axiosPromise("delete", `/reviews/${reviewId}`);
+  };
   return (
     <>
       <div className="mypage-inquire-box-long-wrapper">
@@ -60,18 +28,13 @@ const MyReview = () => {
         <div className="mypage-inquire-box-long">
           {reviewList.map(data => (
             <HandleReview
-              // key={}
+              key={data.reviewsId}
               data={data}
-              // onClickDel={deleteReview}
+              onClickDel={deleteReview}
             />
           ))}
           <div className="return-book-table__pagination">
-            <Pagination
-              page={page}
-              setPage={setPage}
-              lastPage={lastPage}
-              count="10"
-            />
+            <Pagination page={page} setPage={setPage} lastPage={lastPage} />
           </div>
         </div>
       </div>
